# Turns out that passing dom::element by reference can be a performance killer.