Removing one character in the code is enough to boost our API traversal performance.
```
$ ./benchmark/parseandstatcompetition ../jsonexamples/twitter.json
API traversal tests
Based on https://github.com/miloyip/nativejson-benchmark
simdjson (old)                        :   0.454 cycles/byte 	  1.398 instructions/byte 	  7.483 GB/s 11850.026 documents/s
simdjson                                	:   0.360 cycles/byte 	  1.242 instructions/byte 	  9.447 GB/s 14959.982 documents/s
rapid                                   	:   0.471 cycles/byte 	  0.926 instructions/byte 	  7.212 GB/s 11420.218 documents/s

$ ./benchmark/parseandstatcompetition ../jsonexamples/gsoc-2018.json
API traversal tests
Based on https://github.com/miloyip/nativejson-benchmark
simdjson (old)                     	:   0.111 cycles/byte 	  0.368 instructions/byte 	 30.673 GB/s 9217.185 documents/s
simdjson                                	:   0.094 cycles/byte 	  0.327 instructions/byte 	 36.211 GB/s 10881.156 documents/s
rapid                                   	:   0.104 cycles/byte 	  0.257 instructions/byte 	 32.600 GB/s 9796.046 documents/s

$ ./benchmark/parseandstatcompetition ../jsonexamples/random.json
API traversal tests
Based on https://github.com/miloyip/nativejson-benchmark
simdjson  (old)                       :   0.819 cycles/byte 	  3.032 instructions/byte 	  4.145 GB/s 8120.113 documents/s
simdjson                                	:   0.684 cycles/byte 	  2.679 instructions/byte 	  4.963 GB/s 9721.669 documents/s
rapid                                   	:   0.830 cycles/byte 	  2.017 instructions/byte 	  4.091 GB/s 8014.490 documents/s

$ ./benchmark/parseandstatcompetition ../jsonexamples/canada.json
API traversal tests
Based on https://github.com/miloyip/nativejson-benchmark
simdjson   (old)                      :   1.248 cycles/byte 	  4.779 instructions/byte 	  2.719 GB/s 1207.711 documents/s
simdjson                                	:   1.001 cycles/byte 	  4.086 instructions/byte 	  3.390 GB/s 1506.149 documents/s
rapid                                   	:   0.730 cycles/byte 	  2.724 instructions/byte 	  4.649 GB/s 2065.058 documents/s
```


Fixes https://github.com/simdjson/simdjson/issues/932