@lemire 2020-08-01
@jkeiser Can you elaborate on the changes you are requesting?