@jkeiser 2020-08-01
I'm not, just speaking out loud why the reference causes problems :) This seems right to me, merging is good--i haven't had time to go over the documentation or the other PRs yet.